combine scheme modules
